// TRYGO-Front/src/types/MessageType.ts
export enum MessageType {
  ABOUT_LEAN_CANVAS = "ABOUT_LEAN_CANVAS",
  ABOUT_MARKET_RESEARCH = "ABOUT_MARKET_RESEARCH",
  ABOUT_PERSON_PROFILE = "ABOUT_PERSON_PROFILE",
  ABOUT_VALIDATION = 'ABOUT_VALIDATION',
  ABOUT_PACKING = 'ABOUT_PACKING',
  ABOUT_GTM = 'ABOUT_GTM',
  ABOUT_GTM_DETAILED_CHANNEL = 'ABOUT_GTM_DETAILED_CHANNEL',
}